Restrict where classes of pages can be created
